requirements.txt
scipy==1.4.1
numpy==1.18.5

pandas==1.0.4
scikit-learn==0.23.1
tqdm==4.46.1
lightgbm==2.3.1
statsmodels==0.11.1
torch==1.5.0

flask-restful==0.3.8

